ha_ipaper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-ipaper"
version = "1.0.0"
description = "HA-IPaper e-paper dashboard - Home Assistant Add-on"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ha_ipaper"]
